Removed scope methods from LikeableService
